requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.2.7
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
blis==0.4.1
cachetools==4.2.1
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
colorama==0.4.4
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
docopt==0.6.2
en-core-web-sm==2.3.1
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.13
idna==2.9
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
lxml==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.1
opencv-contrib-python==4.4.0.42
packaging==20.4
pandas==1.2.2
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.1.0
pipreqs==0.4.10
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.1
pygame==1.9.6
Pygments==2.7.2
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2019.3
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
soupsieve==2.1
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
streamlit==0.76.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.1
toml==0.10.2
toolz==0.11.1
tornado==6.0.4
tqdm==4.48.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.25.8
validators==0.18.2
wasabi==0.7.1
watchdog==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
yarg==0.1.9
